Move more extension serialization code to x509::extensions.
